sa3_control: bf16 trainer + generate.py + launch + recipes — full pipeline validated
bf16 training (40-step run clean, 222MB ckpt), generate.py (reference-conditioned inference, 50 steps @14 it/s -> 10s riff), launch_riffer.sh (bf16, CK-FA-ready, TunableOp off), recipes/inference_recipes.yaml (9 validated recipes: separation/riffer/steering/guidance/eval) + README. End-to-end train->checkpoint->generate validated vs medium-base.

+++ avp_sa3/sa3_control/conditioner.py
@@ -31,6 +31,6 @@ class AudioRefEncoder(nn.Module):
         self.pool = nn.AdaptiveAvgPool1d(n_tokens)
 
     def forward(self, ref_latent):                 # (B, latent_dim, T)
-        h = self.net(ref_latent.float())           # (B, control_dim, ~T/16)
+        h = self.net(ref_latent.to(self.net[0].weight.dtype))   # (B, control_dim, ~T/16)
         h = self.pool(h)                           # (B, control_dim, n_tokens)
         return h.transpose(1, 2).contiguous()      # (B, n_tokens, control_dim)
